+++ near-contract-standards/src/fungible_token/mod.rs
@@ -7,5 +7,8 @@ pub mod receiver;
 pub mod resolver;
 pub mod storage_impl;
 
+pub use self::core::FungibleTokenCore;
+pub use crate::storage_management::StorageManagement;
 pub use core_impl::FungibleToken;
 pub use macros::*;
+pub use resolver::FungibleTokenResolver;